Giving up on angluar-cookies, switching to jquery-cookie.
angular-cookies doesn't let you specify the path or attributes of the
cookies.

Cookies were being stored under the /list path, and therefore not being
included in requests to /api paths. This meant using CORS to get auth
working, which was also a horrible pain. jquery's cookie library has
more or less the same API, but is just more flexible.